refactor SimpleConfigurationProvider tests
